requirements.txt
docopt==0.6.2
pycryptodome==3.9.8
